Clear detekt thresholds blocking validate by splitting finalisation helpers.
Move lease-record and push helpers into dedicated files, reduce push return count, and apply Spotless to the artifacts fingerprint helper so runtime check passes.
