+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/featuretask/FeatureTaskRuntimeSubtaskFinalisationExtras.kt
@@ -1,9 +1,7 @@
 package skillbill.application.featuretask
 
-import skillbill.ports.workflow.gitops.WorkflowGitRemoteOperations
 import skillbill.ports.workflow.gitops.captureIndexState
 import skillbill.ports.workflow.gitops.headCommitMessage
-import skillbill.ports.workflow.gitops.model.WorkflowGitOperationResult
 import skillbill.ports.workflow.gitops.restoreIndexState
 import skillbill.ports.workflow.gitops.stagePaths
 
@@ -130,9 +128,7 @@ internal fun FeatureTaskRuntimeSubtaskFinalisation.decide(
 internal fun FeatureTaskRuntimeSubtaskFinalisation.headMessage(): String? =
   gitOperations.headCommitMessage(repoRoot).takeIf { it.ok }?.value
 
-internal fun FeatureTaskRuntimeSubtaskFinalisation.ownedHeadAlreadyFinalised(
-  durableCommitSha: String?,
-): Boolean {
+internal fun FeatureTaskRuntimeSubtaskFinalisation.ownedHeadAlreadyFinalised(durableCommitSha: String?): Boolean {
   val durable = durableCommitSha?.trim()?.takeIf(String::isNotBlank) ?: return false
   val headSha = gitOperations.headCommitSha(repoRoot).takeIf { it.ok }?.value?.trim()?.takeIf(String::isNotBlank)
     ?: return false
@@ -145,51 +141,6 @@ internal fun FeatureTaskRuntimeSubtaskFinalisation.remoteDiverged(branch: String
     .takeIf { it.ok }?.value?.trim()?.takeIf(String::isNotBlank) ?: return false
   val ancestor = gitOperations.isCommitAncestor(repoRoot, remoteTip, commitSha)
   return ancestor.ok && ancestor.value.orEmpty().trim().equals("false", ignoreCase = true)
-}
-
-internal fun FeatureTaskRuntimeSubtaskFinalisation.push(
-  branch: String,
-  identity: FeatureTaskRuntimeSubtaskCommitIdentity,
-  commitSha: String,
-  withLease: Boolean,
-): String? {
-  if (!withLease) {
-    val pushed = gitOperations.pushBranch(repoRoot, branch)
-    return if (pushed.ok) null else "the finalised subtask commit '$commitSha' could not be pushed (${pushed.error})"
-  }
-  record(forceWithLeaseRecord(identity, branch, commitSha))
-  val refreshed = refreshRemote(identity, branch)
-  if (refreshed.namesAbsentRemote()) return pushAmended(branch, commitSha)
-  val first = gitOperations.pushBranchWithLease(repoRoot, branch)
-  if (first.ok) return null
-  record(leaseRetryRecord(identity, branch, first.error))
-  val retried = refreshRemote(identity, branch)
-  if (retried.namesAbsentRemote()) return pushAmended(branch, commitSha)
-  val second = gitOperations.pushBranchWithLease(repoRoot, branch)
-  if (second.ok) return null
-  record(leaseAbortRecord(identity, branch, second.error))
-  return "the reopened subtask's amended commit '$commitSha' could not be pushed (${second.error})"
-}
-
-private fun FeatureTaskRuntimeSubtaskFinalisation.pushAmended(branch: String, commitSha: String): String? {
-  val pushed = gitOperations.pushBranch(repoRoot, branch)
-  return if (pushed.ok) null else "the reopened subtask's amended commit '$commitSha' could not be pushed (${pushed.error})"
-}
-
-private fun WorkflowGitOperationResult.namesAbsentRemote(): Boolean =
-  ok && value.trim() == WorkflowGitRemoteOperations.ABSENT_REMOTE_BRANCH
-
-private fun FeatureTaskRuntimeSubtaskFinalisation.refreshRemote(
-  identity: FeatureTaskRuntimeSubtaskCommitIdentity,
-  branch: String,
-): WorkflowGitOperationResult {
-  val refreshed = gitOperations.refreshRemoteBranch(repoRoot, branch)
-  when {
-    refreshed.namesAbsentRemote() -> record(leaseAbsentRecord(identity, branch))
-    refreshed.ok -> record(leaseRefreshRecord(identity, branch))
-    else -> record(leaseRefreshFailedRecord(identity, branch, refreshed.error))
-  }
-  return refreshed
 }
 
 internal fun FeatureTaskRuntimeSubtaskFinalisation.restoring(
