+++ runtime-kotlin/runtime-application/src/main/kotlin/skillbill/application/goalrunner/AttemptLedgerWorkflowDecoding.kt
@@ -18,8 +18,7 @@ internal fun WorkflowStateSnapshot.progressToken(): String = listOf(
   finishedAt.orEmpty(),
 ).joinToString("\n")
 
-internal fun artifactsFingerprint(artifactsJson: String): String =
-  "${artifactsJson.length}:${artifactsJson.hashCode()}"
+internal fun artifactsFingerprint(artifactsJson: String): String = "${artifactsJson.length}:${artifactsJson.hashCode()}"
 
 internal fun decodeWorkflowSteps(stepsJson: String): List<WorkflowStepState> =
   parseWorkflowStepsArray(stepsJson).mapIndexed(::decodeWorkflowStepAt)
